use attribute for test-specific vars
In mchx_pulley_client, we create /site with group "deploy", which
doesn't exist in our kitchen tests, unless we create it first.  So
instead, we just offer an override for the deploy_group attribute
in the tests (adding it to each platform, since that seems to be the
only way that works outside of adding it to the run_list).